+++ src/Fermata.cpp
@@ -348,7 +348,7 @@ struct FermataTextField : STTextField {
   std::unordered_map<int, std::pair<int, int>> lines_to_font_size_and_offset;
 
   FermataTextField() {
-    for (int index = 0; index < 13; index++) {
+    for (int index = 0; index < LARGER_TEXT_COUNT; index++) {
       lines_to_font_size_and_offset.insert({LARGER_TEXT_INFO[index][0],
          std::make_pair(LARGER_TEXT_INFO[index][1], LARGER_TEXT_INFO[index][2])});
     }
@@ -494,8 +494,9 @@ struct FermataWidget : ModuleWidget {
 
   FermataWidget(Fermata* module) {
     setModule(module);
-    setPanel(createPanel(asset::plugin(pluginInstance, "res/Fermata.svg"),
-                         asset::plugin(pluginInstance, "res/Fermata-dark.svg")));
+    setPanel(createPanel<StochasticTelegraph::RestrictedThemedSvgPanel>(
+      asset::plugin(pluginInstance, "res/Fermata.svg"),
+      asset::plugin(pluginInstance, "res/Fermata-dark.svg")));
 
     // Set reasonable initial size of module. Will likely get updated below.
     box.size = Vec(RACK_GRID_WIDTH * Fermata::DEFAULT_WIDTH, RACK_GRID_HEIGHT);
@@ -630,7 +631,7 @@ struct FermataWidget : ModuleWidget {
 
     MenuItem* visible_lines_menu = createSubmenuItem("Visible Lines", "",
       [=](Menu* menu) {
-          for (int index = 0; index < 13; index++) {
+          for (int index = 0; index < LARGER_TEXT_COUNT; index++) {
             int lines = LARGER_TEXT_INFO[index][0];
             menu->addChild(createCheckMenuItem(std::to_string(lines), "",
                 [=]() {return lines == module->visible_lines;},